// src/modules/expense_quotation/enums/expensquotation-status.enum.ts
/* eslint-disable prettier/prettier */
export enum EXPENSQUOTATION_STATUS {
  Nonexistent = 'quotation.status.non_existent',
  Expired = 'quotation.status.expired',
  Draft = 'quotation.status.draft',
  Validated = 'quotation.status.validated',
  Sent = 'quotation.status.sent',
  Accepted = 'quotation.status.accepted',
  Rejected = 'quotation.status.rejected',
  Invoiced = 'quotation.status.invoiced',
  Archived = 'quotation.status.archived',
}